do not add filter word when has trigger_words
